fix: color not generating properly
